third_party/WebKit/LayoutTests/fast/js/Promise-already-rejected.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

window.jsTestIsAsync = true;

new Promise(function(resolve, reject) {
  reject('foo');
  resolve('resolve');
  reject('reject');
}).then(function() {
  testFailed('fulfilled');
  finishJSTest();
}, function(result) {
  window.result = result;
  shouldBeEqualToString('result', 'foo');
  finishJSTest();
});

</script>
</body>
</html>
